requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=pypi_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=pypi_0
chardet=3.0.4=pypi_0
charset-normalizer=3.1.0=pypi_0
comm=0.1.2=py39h06a4308_0
contourpy=1.1.0=pypi_0
cryptography=41.0.1=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
docopt=0.6.2=pypi_0
dwdweather2=0.14.0=pypi_0
entrypoints=0.4=py39h06a4308_0
entsoe-py=0.5.9=pypi_0
executing=0.8.3=pyhd3eb1b0_0
flatbuffers=23.5.26=pypi_0
fonttools=4.40.0=pypi_0
frozendict=2.3.8=pypi_0
fusepy=3.0.1=pypi_0
gast=0.4.0=pypi_0
google-auth=2.20.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.2=pypi_0
h5py=3.9.0=pypi_0
html5lib=1.1=pypi_0
htmllistparse=0.5.2=pypi_0
idna=2.8=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
ipykernel=6.19.2=py39hb070fc8_0
ipython=8.8.0=py39h06a4308_0
jax=0.4.12=pypi_0
jedi=0.18.1=py39h06a4308_1
joblib=1.2.0=pypi_0
jupyter_client=7.4.8=py39h06a4308_0
jupyter_core=5.1.1=py39h06a4308_0
keras=2.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.0=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
lightgbm=3.3.5=pypi_0
lxml=4.9.2=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=py39h06a4308_0
missingno=0.5.2=pypi_0
ml-dtypes=0.2.0=pypi_0
multiprocessing-logging=0.3.4=pypi_0
multitasking=0.0.11=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py39h06a4308_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.8=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=22.0=py39h06a4308_0
pandas=2.0.2=pypi_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=23.1.2=py39h06a4308_0
pipreqs=0.4.13=pypi_0
platformdirs=2.5.2=py39h06a4308_0
prompt-toolkit=3.0.36=py39h06a4308_0
protobuf=4.23.3=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.1.0=pypi_0
python=3.9.16=h955ad1f_3
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2023.3=pypi_0
pyzmq=23.2.0=py39h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-cache=0.5.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.8.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.14.0=pypi_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.32.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tornado=6.2=py39h5eee18b_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py39h06a4308_0
typing-extensions=4.6.3=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.25.11=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py39h06a4308_0
wrapt=1.14.1=pypi_0
xgboost=1.7.6=pypi_0
xz=5.4.2=h5eee18b_0
yahoofinancials=1.15=pypi_0
yarg=0.1.9=pypi_0
yfinance=0.2.20=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
